requirements.txt
asgiref==3.7.2
Babel==2.14.0
bleach==6.1.0
Django==4.2.9
django-appconf==1.0.6
django-cors-headers==4.3.1
django-imagekit==5.0.0
django-money==3.4.1
django-preventconcurrentlogins==0.8.2
django-summernote==0.8.20.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pilkit==3.0
pillow==10.2.0
py-moneyed==3.0
PyJWT==2.8.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
webencodings==0.5.1
